Download additional data collected by annee politique suisse (the complete SwissVotes-Database)

get_swissvotes downloads additional data collected by annee politique suisse. It allows for completely downloading their database. Please cite data.

get_swissvotes(DB = T, savecitation = F, codebook = F)

Arguments

DB
get database

savecitation
by default = FALSE. Saves the citation within a .txt file in the working directory if TRUE.

codebook
by default = FALSE. If TRUE navigates your browser to the codebook.

Value

a tibble containing the results

Details

get_swissvotes - retrieve data on votes. The unit of analysis are votes.

Examples

 # \donttest{

# See codebook only
 get_swissvotes(codebook=F)

#> To cite swissvotes data in publications, please use:
#> 
#>             Linder, Wolf, Christian Bolliger und Yvan Rielle (2010): 
#>             Verzeichnisse der Literatur, der Quellen und der Abkuerzungen fuer die Kurzbeschreibungen zu den Abstimmungen 1848-2007. 
#>             In: Linder, Wolf, Christian Bolliger und Yvan Rielle (Hg.): Handbuch der eidgenoessischen Volksabstimmungen 1848-2007. Bern: Haupt. S. 713-729.
#> 
#>             
#>             A BibTeX entry for LaTeX users is
#> 
#>             @Book{,
#>     author = {Linder, Wolf and Bolliger, Christian and Rielle, Yvan},
#>             title = {Handbuch der eidgenoessischen Volksabstimmungen 1848-2007},
#>             publisher = {Haupt, Bern},
#>             year = {2010},
#>             url = {https://swissvotes.ch},
#>   }

#>      anr datum titel_kurz_d titel_kurz_f titel_off_d titel_off_f stichwort
#>      anzahl rechtsform d1e1 d1e2 d1e3 d2e1 d2e2 d2e3 d3e1 d3e2 d3e3 volk stand
#>      annahme berecht stimmen bet leer ungultig gultig volkja volknein
#>      volkja.proz kt.ja kt.nein ktjaproz zh.berecht zh.stimmen zh.bet zh.gultig
#>      zh.ja zh.nein zh.japroz zh.annahme be.berecht be.stimmen be.bet be.gultig
#>      be.ja be.nein be.japroz be.annahme lu.berecht lu.stimmen lu.bet lu.gultig
#>      lu.ja lu.nein lu.japroz lu.annahme ur.berecht ur.stimmen ur.bet ur.gultig
#>      ur.ja ur.nein ur.japroz ur.annahme sz.berecht sz.stimmen sz.bet sz.gultig
#>      sz.ja sz.nein sz.japroz sz.annahme ow.berecht ow.stimmen ow.bet ow.gultig
#>      ow.ja ow.nein ow.japroz ow.annahme nw.berecht nw.stimmen nw.bet nw.gultig
#>      nw.ja nw.nein nw.japroz nw.annahme gl.berecht gl.stimmen gl.bet gl.gultig
#>      gl.ja gl.nein gl.japroz gl.annahme zg.berecht zg.stimmen zg.bet zg.gultig
#>      zg.ja zg.nein zg.japroz zg.annahme fr.berecht fr.stimmen fr.bet fr.gultig
#>      fr.ja fr.nein fr.japroz fr.annahme so.berecht so.stimmen so.bet so.gultig
#>      so.ja so.nein so.japroz so.annahme bs.berecht bs.stimmen bs.bet bs.gultig
#>      bs.ja bs.nein bs.japroz bs.annahme bl.berecht bl.stimmen bl.bet bl.gultig
#>      bl.ja bl.nein bl.japroz bl.annahme sh.berecht sh.stimmen sh.bet sh.gultig
#>      sh.ja sh.nein sh.japroz sh.annahme ar.berecht ar.stimmen ar.bet ar.gultig
#>      ar.ja ar.nein ar.japroz ar.annahme ai.berecht ai.stimmen ai.bet ai.gultig
#>      ai.ja ai.nein ai.japroz ai.annahme sg.berecht sg.stimmen sg.bet sg.gultig
#>      sg.ja sg.nein sg.japroz sg.annahme gr.berecht gr.stimmen gr.bet gr.gultig
#>      gr.ja gr.nein gr.japroz gr.annahme ag.berecht ag.stimmen ag.bet ag.gultig
#>      ag.ja ag.nein ag.japroz ag.annahme tg.berecht tg.stimmen tg.bet tg.gultig
#>      tg.ja tg.nein tg.japroz tg.annahme ti.berecht ti.stimmen ti.bet ti.gultig
#>      ti.ja ti.nein ti.japroz ti.annahme vd.berecht vd.stimmen vd.bet vd.gultig
#>      vd.ja vd.nein vd.japroz vd.annahme vs.berecht vs.stimmen vs.bet vs.gultig
#>      vs.ja vs.nein vs.japroz vs.annahme ne.berecht ne.stimmen ne.bet ne.gultig
#>      ne.ja ne.nein ne.japroz ne.annahme ge.berecht ge.stimmen ge.bet ge.gultig
#>      ge.ja ge.nein ge.japroz ge.annahme ju.berecht ju.stimmen ju.bet ju.gultig
#>      ju.ja ju.nein ju.japroz ju.annahme dep br.pos legislatur legisjahr
#>      gesch_nr bv.pos nr.pos nrja nrnein sr.pos srja srnein dauer_bv dauer_abst
#>      i.dauer_samm i.dauer_br i.dauer_tot fr.dauer_samm fr.dauer_tot unter_g
#>      unter_u urheber p.fdp p.cvp p.sps p.svp p.evp p.ucsp p.pda p.gps p.sd
#>      p.edu p.fps p.lega p.kvp p.glp p.bdp p.mcg p.lps p.ldu p.poch p.rep p.sav
#>      p.eco p.sgv p.sbv p.sgb p.travs p.vsa p.vpod p.voev p.tcs p.vcs p.acs
#>      p.sbk p.ssv p.gem p.kdk p.kkjpd p.gdk p.ldk p.vdk p.sodk p.endk p.fdk
#>      p.edk p.bpuk p.others_yes p.others_no p.others_free pdev.bdp_AG
#>      pdev.bdp_AI pdev.bdp_AR pdev.bdp_BE pdev.bdp_BL pdev.bdp_BS pdev.bdp_FR
#>      pdev.bdp_GE pdev.bdp_GL pdev.bdp_GR pdev.bdp_JU pdev.bdp_LU pdev.bdp_NE
#>      pdev.bdp_NW pdev.bdp_OW pdev.bdp_SG pdev.bdp_SH pdev.bdp_SO pdev.bdp_SZ
#>      pdev.bdp_TG pdev.bdp_TI pdev.bdp_UR pdev.bdp_VD pdev.bdp_VS pdev.bdp_VSr
#>      pdev.bdp_VSo pdev.bdp_ZG pdev.bdp_ZH pdev.jbdp_CH pdev.csp_FR pdev.csp_GR
#>      pdev.csp_JU pdev.csp_LU pdev.csp_OW pdev.csp_SG pdev.csp_VS pdev.csp_VSr
#>      pdev.csp_VSo pdev.csp_ZH pdev.cvp_frauen pdev.cvp_AG pdev.cvp_AI
#>      pdev.cvp_AR pdev.cvp_BE pdev.cvp_BL pdev.cvp_BS pdev.cvp_FR pdev.cvp_GE
#>      pdev.cvp_GL pdev.cvp_GR pdev.cvp_JU pdev.cvp_LU pdev.cvp_NE pdev.cvp_NW
#>      pdev.cvp_OW pdev.cvp_SG pdev.cvp_SH pdev.cvp_SO pdev.cvp_SZ pdev.cvp_TG
#>      pdev.cvp_TI pdev.cvp_UR pdev.cvp_VD pdev.cvp_VS pdev.cvp_VSr pdev.cvp_VSo
#>      pdev.cvp_ZG pdev.cvp_ZH pdev.jcvp_CH pdev.jcvp_AG pdev.jcvp_BE
#>      pdev.jcvp_GR pdev.jcvp_LU pdev.jcvp_SO pdev.jcvp_ZH pdev.edu_AG
#>      pdev.edu_AI pdev.edu_AR pdev.edu_BE pdev.edu_BL pdev.edu_BS pdev.edu_FR
#>      pdev.edu_GE pdev.edu_GL pdev.edu_GR pdev.edu_JU pdev.edu_LU pdev.edu_NE
#>      pdev.edu_NW pdev.edu_OW pdev.edu_SG pdev.edu_SH pdev.edu_SO pdev.edu_SZ
#>      pdev.edu_TG pdev.edu_TI pdev.edu_UR pdev.edu_VD pdev.edu_VS pdev.edu_VSr
#>      pdev.edu_VSo pdev.edu_ZG pdev.edu_ZH pdev.evp_AG pdev.evp_AI pdev.evp_AR
#>      pdev.evp_BE pdev.evp_BL pdev.evp_BS pdev.evp_FR pdev.evp_GE pdev.evp_GL
#>      pdev.evp_GR pdev.evp_JU pdev.evp_LU pdev.evp_NE pdev.evp_NW pdev.evp_OW
#>      pdev.evp_SG pdev.evp_SH pdev.evp_SO pdev.evp_SZ pdev.evp_TG pdev.evp_TI
#>      pdev.evp_UR pdev.evp_VD pdev.evp_VS pdev.evp_ZG pdev.evp_ZH pdev.jevp_CH
#>      pdev.fdp_Frauen pdev.fdp_AG pdev.fdp_AI pdev.fdp_AR pdev.fdp_BE
#>      pdev.fdp_BL pdev.fdp_BS pdev.fdp_FR pdev.fdp_GE pdev.fdp_GL pdev.fdp_GR
#>      pdev.fdp_JU pdev.fdp_LU pdev.fdp_NE pdev.fdp_NW pdev.fdp_OW pdev.fdp_SG
#>      pdev.fdp_SH pdev.fdp_SO pdev.fdp_SZ pdev.fdp_TG pdev.fdp_TI pdev.fdp_UR
#>      pdev.fdp_VD pdev.fdp_VS pdev.fdp_VSr pdev.fdp_Vso pdev.fdp_ZG pdev.fdp_ZH
#>      pdev.jfdp_CH pdev.jfdp_AG pdev.jfdp_BL pdev.jfdp_FR pdev.jfdp_GR
#>      pdev.jfdp_JU pdev.jfdp_LU pdev.jfdp_SH pdev.jfdp_TI pdev.jfdp_VD
#>      pdev.jfdp_ZH pdev.fps_AG pdev.fps_AI pdev.fps_BE pdev.fps_BL pdev.fps_BS
#>      pdev.fps_SG pdev.fps_SH pdev.fps_SO pdev.fps_TG pdev.fps_ZH pdev.glp_AG
#>      pdev.glp_AI pdev.glp_AR pdev.glp_BE pdev.glp_BL pdev.glp_BS pdev.glp_FR
#>      pdev.glp_GE pdev.glp_GL pdev.glp_GR pdev.glp_JU pdev.glp_LU pdev.glp_NE
#>      pdev.glp_NW pdev.glp_OW pdev.glp_SG pdev.glp_SH pdev.glp_SO pdev.glp_SZ
#>      pdev.glp_TG pdev.glp_TI pdev.glp_UR pdev.glp_VD pdev.glp_VS pdev.glp_VSr
#>      pdev.glp_VSo pdev.glp_ZG pdev.glp_ZH pdev.jglp_CH pdev.gps_AG pdev.gps_AI
#>      pdev.gps_AR pdev.gps_BE pdev.gps_BL pdev.gps_BS pdev.gps_FR pdev.gps_GE
#>      pdev.gps_GL pdev.gps_GR pdev.gps_JU pdev.gps_LU pdev.gps_NE pdev.gps_NW
#>      pdev.gps_OW pdev.gps_SG pdev.gps_SH pdev.gps_SO pdev.gps_SZ pdev.gps_TG
#>      pdev.gps_TI pdev.gps_UR pdev.gps_VD pdev.gps_VS pdev.gps_VSr pdev.gps_VSo
#>      pdev.gps_ZG pdev.gps_ZH pdev.jgps_CH pdev.kvp_SG pdev.lps_BE pdev.lps_BL
#>      pdev.lps_BS pdev.lps_FR pdev.lps_GE pdev.lps_JU pdev.lps_NE pdev.lps_SG
#>      pdev.lps_TI pdev.lps_VD pdev.lps_VS pdev.lps_ZH pdev.jlps_CH pdev.jlps_SO
#>      pdev.jlps_TI pdev.ldu_AG pdev.ldu_AR pdev.ldu_BE pdev.ldu_BL pdev.ldu_BS
#>      pdev.ldu_GR pdev.ldu_LU pdev.ldu_NE pdev.ldu_SG pdev.ldu_SH pdev.ldu_SO
#>      pdev.ldu_TG pdev.ldu_VD pdev.ldu_ZG pdev.ldu_ZH pdev.jldu_CH pdev.poch_SO
#>      pdev.poch_ZH pdev.pda_BE pdev.pda_BL pdev.pda_BS pdev.pda_GE pdev.pda_JU
#>      pdev.pda_NE pdev.pda_SG pdev.pda_TI pdev.pda_VD pdev.pda_ZH pdev.jpda_CH
#>      pdev.rep_AG pdev.rep_GE pdev.rep_NE pdev.rep_TG pdev.rep_VD pdev.rep_ZH
#>      pdev.sd_AG pdev.sd_BE pdev.sd_BL pdev.sd_BS pdev.sd_FR pdev.sd_GE
#>      pdev.sd_GR pdev.sd_LU pdev.sd_NE pdev.sd_SG pdev.sd_SO pdev.sd_TG
#>      pdev.sd_TI pdev.sd_VD pdev.sd_ZH pdev.jsd_CH pdev.sps_AG pdev.sps_AI
#>      pdev.sps_AR pdev.sps_BE pdev.sps_BL pdev.sps_BS pdev.sps_FR pdev.sps_GE
#>      pdev.sps_GL pdev.sps_GR pdev.sps_JU pdev.sps_LU pdev.sps_NE pdev.sps_NW
#>      pdev.sps_OW pdev.sps_SG pdev.sps_SH pdev.sps_SO pdev.sps_SZ pdev.sps_TG
#>      pdev.sps_TI pdev.sps_UR pdev.sps_VD pdev.sps_VS pdev.sps_VSr pdev.sps_VSo
#>      pdev.sps_ZG pdev.sps_ZH pdev.juso_CH pdev.juso_BE pdev.juso_GE
#>      pdev.juso_JU pdev.juso_TI pdev.juso_VS pdev.juso_ZH pdev.svp_AG
#>      pdev.svp_AI pdev.svp_AR pdev.svp_BE pdev.svp_BL pdev.svp_BS pdev.svp_FR
#>      pdev.svp_GE pdev.svp_GL pdev.svp_GR pdev.svp_JU pdev.svp_LU pdev.svp_NE
#>      pdev.svp_NW pdev.svp_OW pdev.svp_SG pdev.svp_SH pdev.svp_SO pdev.svp_SZ
#>      pdev.svp_TG pdev.svp_TI pdev.svp_UR pdev.svp_VD pdev.svp_VS pdev.svp_VSr
#>      pdev.svp_VSo pdev.svp_ZG pdev.svp_ZH pdev.jsvp_CH pdev.jsvp_AG
#>      pdev.jsvp_BE pdev.jsvp_GE pdev.jsvp_SH pdev.jsvp_UR pdev.jsvp_ZH
#>      pdev.sgb_AG pdev.sgb_JU pdev.sgb_VS pdev.sgv_AG pdev.sgv_BS pdev.sgv_SH
#>      pdev.vpod_GE pdev.vpod_VD nr.wahl w.fdp w.cvp w.sp w.svp w.evp w.csp w.pda
#>      w.gps w.sd w.edu w.fps w.lega w.kvp w.glp w.bdp w.mcg w.lps w.ldu w.poch
#>      w.rep w.ubrige ja.lager nein.lager keinepar.summe leer.summe
#>      freigabe.summe neutral.summe unbekannt.summe anneepolitique bfsmap.de
#>      bfsmap.fr curiavista.de curiavista.fr bkresults.de bkresults.fr
#>      bkchrono.de bkchrono.fr X X.1 X.2 X.3 X.4 X.5 X.6 X.7 X.8 X.9 X.10 X.11
#>      X.12 X.13 X.14 X.15 X.16 X.17 X.18 X.19 X.20 X.21 X.22 X.23 X.24 X.25 X.26
#>      X.27 X.28 X.29 X.30 X.31 X.32 X.33 X.34 X.35 X.36 X.37 X.38 X.39 X.40 X.41
#>      X.42 X.43 X.44 X.45 X.46 X.47 X.48 X.49 X.50 X.51 X.52 X.53 X.54 X.55 X.56
#>      X.57 X.58 X.59 X.60 X.61 X.62 X.63 X.64 X.65 X.66 X.67 X.68 X.69 X.70 X.71
#>      X.72 X.73 X.74 X.75 X.76 X.77 X.78 X.79 X.80 X.81 X.82 X.83 X.84 X.85 X.86
#>      X.87 X.88 X.89 X.90 X.91 X.92 X.93 X.94 X.95 X.96 X.97 X.98 X.99 X.100
#>      X.101 X.102 X.103 X.104 X.105 X.106 X.107 X.108 X.109 X.110 X.111 X.112
#>      X.113 X.114 X.115 X.116 X.117 X.118 X.119 X.120 X.121 X.122 X.123 X.124
#>      X.125 X.126 X.127 X.128 X.129 X.130 X.131 X.132 X.133 X.134 X.135 X.136
#>      X.137 X.138 X.139 X.140 X.141 X.142 X.143 X.144 X.145 X.146 X.147 X.148
#>      X.149 X.150 X.151 X.152 X.153 X.154 X.155 X.156 X.157 X.158 X.159 X.160
#>      X.161 X.162 X.163 X.164 X.165 X.166 X.167 X.168 X.169 X.170 X.171 X.172
#>      X.173 X.174 X.175 X.176 X.177 X.178 X.179 X.180 X.181 X.182 X.183 X.184
#>      X.185 X.186 X.187 X.188 X.189 X.190 X.191 X.192 X.193 X.194 X.195 X.196
#>      X.197 X.198 X.199 X.200 X.201 X.202 X.203 X.204 X.205 X.206 X.207 X.208
#>      X.209 X.210 X.211 X.212 X.213 X.214 X.215 X.216 X.217 X.218 X.219 X.220
#>      X.221 X.222 X.223 X.224 X.225 X.226 X.227 X.228 X.229 X.230 X.231 X.232
#>      X.233 X.234 X.235 X.236 X.237 X.238 X.239 X.240 X.241 X.242 X.243 X.244
#>      X.245 X.246 X.247 X.248 X.249 X.250 X.251 X.252 X.253 X.254 X.255 X.256
#>      X.257 X.258 X.259 X.260 X.261 X.262 X.263 X.264 X.265 X.266 X.267 X.268
#>      X.269 X.270 X.271 X.272 X.273 X.274 X.275 X.276 X.277 X.278 X.279 X.280
#>      X.281 X.282 X.283 X.284 X.285 X.286 X.287 X.288 X.289 X.290 X.291 X.292
#>      X.293 X.294 X.295 X.296
#>  [ reached 'max' / getOption("max.print") -- omitted 649 rows ]

glimpse(results)

#> Error in glimpse(results): could not find function "glimpse"

# }