Transform cantonal results json into tibble

canton_json_to_dfr Tranforms a single results json for a selected cantonal votedate into a tibble.

canton_json_to_dfr(votedate = NULL, geolevel = "municipality",
  dataurl = NULL, index = NULL)

Arguments

votedate
date of the ballot. Default: most recent ballot available.

geolevel
geographical level for which the results should be loaded. options."canton","district" or "municipality"

dataurl
list of datasets / metadata for the given dataset and its resources OR url of the dcat dataset on opendata.swiss

index
selection by index of the resource (last published = 1).

Value

a tibble containing the results

Details

placeholder

Examples

 # \donttest{

#most recent vote
results <- canton_json_to_dfr()

glimpse(results)

#> Error in glimpse(results): could not find function "glimpse"

# transform the json for a single votedate at counting district level
canton_json_to_dfr(votedate="2019-09-22",geolevel = "zh_counting_districts")

#> # A tibble: 33 x 17
#>        id canton_name mun_id mun_name district_id gebietAusgezaeh…
#>     <int> <chr>       <chr>  <chr>    <chr>       <lgl>           
#>  1 104897 OW          1401   Alpnach  600         TRUE            
#>  2 104897 OW          1402   Engelbe… 600         TRUE            
#>  3 104897 OW          1403   Giswil   600         TRUE            
#>  4 104897 OW          1404   Kerns    600         TRUE            
#>  5 104897 OW          1405   Lungern  600         TRUE            
#>  6 104897 OW          1406   Sachseln 600         TRUE            
#>  7 104897 OW          1407   Sarnen   600         TRUE            
#>  8 104898 SH          2901   Gächlin… 1401        TRUE            
#>  9 104898 SH          2903   Löhning… 1401        TRUE            
#> 10 104898 SH          2904   Neunkir… 1401        TRUE            
#> # … with 23 more rows, and 11 more variables: jaStimmenInProzent <dbl>,
#> #   jaStimmenAbsolut <int>, neinStimmenAbsolut <int>,
#> #   stimmbeteiligungInProzent <dbl>, eingelegteStimmzettel <int>,
#> #   anzahlStimmberechtigte <int>, gueltigeStimmen <int>, de <chr>, fr <chr>,
#> #   it <chr>, rm <chr>

# }