Transform a opendata.swiss cantonal results json into a tibble

canton_json_to_dfr Tranforms a single results json for a selected cantonal votedate into a tibble.

canton_json_to_dfr(votedate = NULL, geolevel = "municipality",
  dataurl = NULL, index = NULL)

Arguments

votedate
date of the ballot. Default: most recent ballot available.

geolevel
geographical level for which the results should be loaded. options."canton","district" or "municipality"

dataurl
list of datasets / metadata for the given dataset and its resources OR url of the dcat dataset on opendata.swiss

index
selection by index of the resource (last published = 1).

Value

a tibble containing the results

Examples

# get and transform the json for the most recent vote
results <- canton_json_to_dfr()

# get and transform the json for a single votedate at counting district level
canton_json_to_dfr(votedate="2020-02-09",geolevel = "zh_counting_districts")

#> # A tibble: 2,880 x 17
#>        id canton_name mun_id mun_name geoLevelParentn… gebietAusgezaeh…
#>     <int> <chr>       <chr>  <chr>    <chr>            <lgl>           
#>  1 104947 ZH          1      Aeugst … 101              TRUE            
#>  2 104947 ZH          2      Affolte… 101              TRUE            
#>  3 104947 ZH          3      Bonstet… 101              TRUE            
#>  4 104947 ZH          4      Hausen … 101              TRUE            
#>  5 104947 ZH          5      Hedingen 101              TRUE            
#>  6 104947 ZH          6      Kappel … 101              TRUE            
#>  7 104947 ZH          7      Knonau   101              TRUE            
#>  8 104947 ZH          8      Maschwa… 101              TRUE            
#>  9 104947 ZH          9      Mettmen… 101              TRUE            
#> 10 104947 ZH          10     Obfelden 101              TRUE            
#> # … with 2,870 more rows, and 11 more variables: jaStimmenInProzent <dbl>,
#> #   jaStimmenAbsolut <int>, neinStimmenAbsolut <int>,
#> #   stimmbeteiligungInProzent <dbl>, eingelegteStimmzettel <int>,
#> #   anzahlStimmberechtigte <int>, gueltigeStimmen <int>, de <chr>, fr <chr>,
#> #   it <chr>, rm <chr>