Get cantonal results and counting status for selected dates or a given period

get_cantonalsvotes is one of the two main functions of swissvote package. It allows to retrieve the results and the counting status for national ballots.

get_cantonalvotes(geolevel = "municipality", votedates = NULL,
  from_date = NULL, to_date = NULL)

Arguments

geolevel
geographical level for which the results should be loaded. options: "canton", "district","municipality" or "zh_counting_districts"

votedates
dates of the ballots to be selected. Default: most recent ballot available. Format = "YYYY-MM-DD"

from_date
starting point in time from which vote-results should be retrived. Format = "YYYY-MM-DD"

to_date
end point in time to which vote-results should be retrived. Format = "YYYY-MM-DD"

Value

a tibble containing the results

Details

get_cantonalvotes - retrieve vote results for cantonal ballots at district- or municipality level for selected dates or a given date range.

Examples

 # \donttest{
results <-get_cantonalvotes(geolevel="district",from_date = 20180101,to_date=20181231)

 get_cantonalvotes(to_date="1983-12-04")

#> # A tibble: 0 x 0

 OR

#> Error in eval(expr, envir, enclos): object 'OR' not found

 get_cantonalvotes(votedates="2019-02-10")

#> # A tibble: 2,341 x 18
#>        id canton_name mun_id mun_name district_id gebietAusgezaeh…
#>     <int> <chr>       <chr>  <chr>    <chr>       <lgl>           
#>  1 104812 ZH          1      Aeugst … 101         TRUE            
#>  2 104812 ZH          2      Affolte… 101         TRUE            
#>  3 104812 ZH          3      Bonstet… 101         TRUE            
#>  4 104812 ZH          4      Hausen … 101         TRUE            
#>  5 104812 ZH          5      Hedingen 101         TRUE            
#>  6 104812 ZH          6      Kappel … 101         TRUE            
#>  7 104812 ZH          7      Knonau   101         TRUE            
#>  8 104812 ZH          8      Maschwa… 101         TRUE            
#>  9 104812 ZH          9      Mettmen… 101         TRUE            
#> 10 104812 ZH          10     Obfelden 101         TRUE            
#> # … with 2,331 more rows, and 12 more variables: jaStimmenInProzent <dbl>,
#> #   jaStimmenAbsolut <int>, neinStimmenAbsolut <int>,
#> #   stimmbeteiligungInProzent <dbl>, eingelegteStimmzettel <int>,
#> #   anzahlStimmberechtigte <int>, gueltigeStimmen <int>, de <chr>, fr <chr>,
#> #   it <chr>, rm <chr>, votedate <chr>

glimpse(results)

#> Error in glimpse(results): could not find function "glimpse"

# get the results at counting district level
# yields the same result as the municipality level, with the exception of Winterthur and Zurich, where detailed counting district results are returned
get_cantonalvotes(votedate="2019-09-22",geolevel = "zh_counting_districts")

#> # A tibble: 33 x 18
#>        id canton_name mun_id mun_name district_id gebietAusgezaeh…
#>     <int> <chr>       <chr>  <chr>    <chr>       <lgl>           
#>  1 104897 OW          1401   Alpnach  600         TRUE            
#>  2 104897 OW          1402   Engelbe… 600         TRUE            
#>  3 104897 OW          1403   Giswil   600         TRUE            
#>  4 104897 OW          1404   Kerns    600         TRUE            
#>  5 104897 OW          1405   Lungern  600         TRUE            
#>  6 104897 OW          1406   Sachseln 600         TRUE            
#>  7 104897 OW          1407   Sarnen   600         TRUE            
#>  8 104898 SH          2901   Gächlin… 1401        TRUE            
#>  9 104898 SH          2903   Löhning… 1401        TRUE            
#> 10 104898 SH          2904   Neunkir… 1401        TRUE            
#> # … with 23 more rows, and 12 more variables: jaStimmenInProzent <dbl>,
#> #   jaStimmenAbsolut <int>, neinStimmenAbsolut <int>,
#> #   stimmbeteiligungInProzent <dbl>, eingelegteStimmzettel <int>,
#> #   anzahlStimmberechtigte <int>, gueltigeStimmen <int>, de <chr>, fr <chr>,
#> #   it <chr>, rm <chr>, votedate <chr>

# }