Get Swiss Geodata

get_geodata retrieves the latest geodata provided by the Federal Statistical Office in connection with federal votes.

get_geodata(geolevel = "municipality", latest = T, verbose = F, call_res)

Arguments

geolevel
geographical level. Options: "national", "canton", "district", "municipality", "zh_counting_districts" or "lakes".

latest
if TRUE, the latest data is retrieved. If FALSE, geo data from the beginning of the year is retrieved. The API does not support finer distinctions. For more detailed information on the exact status of the data, please use verbode = TRUE.

verbose
if TRUE, the date from which the data originates is displayed.

call_res
result of a previous call to the geodata API. Optional argument.

Value

a simple feature collection of the desired spatial units with corresponding IDs.

Examples

# Get latest geodata at municipal level
get_geodata()

#> Simple feature collection with 2192 features and 1 field
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 485203 ymin: 75283.29 xmax: 834035 ymax: 295868.1
#> CRS:            NA
#> First 10 features:
#>    mun_id                       geometry
#> 1       1 MULTIPOLYGON (((678567.2 23...
#> 2       2 MULTIPOLYGON (((673828 2385...
#> 3       3 MULTIPOLYGON (((676433.7 24...
#> 4       4 MULTIPOLYGON (((680613.3 23...
#> 5       5 MULTIPOLYGON (((675419.4 24...
#> 6       6 MULTIPOLYGON (((679214.2 23...
#> 7       7 MULTIPOLYGON (((676328.7 23...
#> 8       8 MULTIPOLYGON (((675436.8 23...
#> 9       9 MULTIPOLYGON (((676171.3 23...
#> 10     10 MULTIPOLYGON (((672866.1 23...

# Get latest geodata at cantonal level
get_geodata(geolevel = "canton")

#> Simple feature collection with 26 features and 1 field
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 485203 ymin: 75283.29 xmax: 834035 ymax: 295868.1
#> CRS:            NA
#> First 10 features:
#>    canton_id                       geometry
#> 1          1 MULTIPOLYGON (((674282.6 26...
#> 2          2 MULTIPOLYGON (((569582.3 20...
#> 3          3 MULTIPOLYGON (((646441.8 18...
#> 4          4 MULTIPOLYGON (((695023.3 15...
#> 5          5 MULTIPOLYGON (((695058.3 22...
#> 6          6 MULTIPOLYGON (((666220.7 20...
#> 7          7 MULTIPOLYGON (((686401.8 20...
#> 8          8 MULTIPOLYGON (((718719.5 22...
#> 9          9 MULTIPOLYGON (((695058.3 22...
#> 10        10 MULTIPOLYGON (((548754.2 18...