Get national results and counting status in real time or for selected dates or a time range in the past

get_nationalvotes is one of the two main functions of swissvote package. It allows to retrieve the results and the counting status for national ballots.

get_nationalvotes(
  geolevel = "municipality",
  votedates = NULL,
  from_date = NULL,
  to_date = NULL,
  language = "DE"
)

Arguments

geolevel
geographical level for which the results should be loaded. Options: "national", "canton", "district", "municipality" or "zh_counting_districts".#' @param votedates dates of the ballots to be selected. Default: most recent ballot available. Format: "YYYY-MM-DD".

votedates
dates of the ballots to be selected. Default: most recent ballot available. Format: "YYYY-MM-DD".

from_date
starting point in time from which vote results should be retrieved. Format: "YYYY-MM-DD".

to_date
end point in time to which vote results should be retrieved. Format: "YYYY-MM-DD".

language
defines the language of the vote title. Options: "DE" for German, "FR" for French, "IT" for Italian or "RM" for Romansh.

Value

a tibble containing the results

Details

get_nationalvotes - retrieve vote results for national ballots at district- or municipality level for selected dates or a given date range.

Examples

# \donttest{
# Selection by range
results <- get_nationalvotes(
    geolevel = "district", 
    from_date = "2018-01-01",
    to_date = "2018-12-31"
    )

# Selection by end date only
get_nationalvotes(to_date = "1983-12-04")

#> # A tibble: 23,678 × 16
#>    name              id canton_id canton_name mun_id mun_name   geoLevelParentn…
#>    <chr>          <int> <chr>     <chr>       <chr>  <chr>      <chr>           
#>  1 Volksinitiati…  3111 1         Zürich      1      Aeugst am… 101             
#>  2 Volksinitiati…  3111 1         Zürich      2      Affoltern… 101             
#>  3 Volksinitiati…  3111 1         Zürich      3      Bonstetten 101             
#>  4 Volksinitiati…  3111 1         Zürich      4      Hausen am… 101             
#>  5 Volksinitiati…  3111 1         Zürich      5      Hedingen   101             
#>  6 Volksinitiati…  3111 1         Zürich      6      Kappel am… 101             
#>  7 Volksinitiati…  3111 1         Zürich      7      Knonau     101             
#>  8 Volksinitiati…  3111 1         Zürich      8      Maschwand… 101             
#>  9 Volksinitiati…  3111 1         Zürich      9      Mettmenst… 101             
#> 10 Volksinitiati…  3111 1         Zürich      10     Obfelden   101             
#> # … with 23,668 more rows, and 9 more variables: gebietAusgezaehlt <lgl>,
#> #   jaStimmenInProzent <dbl>, jaStimmenAbsolut <int>, neinStimmenAbsolut <int>,
#> #   stimmbeteiligungInProzent <dbl>, eingelegteStimmzettel <int>,
#> #   anzahlStimmberechtigte <int>, gueltigeStimmen <int>, votedate <date>

 # }
# Selection of a specific vote date
get_nationalvotes(votedates = "2014-02-09")

#> # A tibble: 6,519 × 16
#>    name               id canton_id canton_name mun_id mun_name  geoLevelParentn…
#>    <chr>           <int> <chr>     <chr>       <chr>  <chr>     <chr>           
#>  1 Bundesbeschlus…  5780 1         Zürich      1      Aeugst a… 101             
#>  2 Bundesbeschlus…  5780 1         Zürich      2      Affolter… 101             
#>  3 Bundesbeschlus…  5780 1         Zürich      3      Bonstett… 101             
#>  4 Bundesbeschlus…  5780 1         Zürich      4      Hausen a… 101             
#>  5 Bundesbeschlus…  5780 1         Zürich      5      Hedingen  101             
#>  6 Bundesbeschlus…  5780 1         Zürich      6      Kappel a… 101             
#>  7 Bundesbeschlus…  5780 1         Zürich      7      Knonau    101             
#>  8 Bundesbeschlus…  5780 1         Zürich      8      Maschwan… 101             
#>  9 Bundesbeschlus…  5780 1         Zürich      9      Mettmens… 101             
#> 10 Bundesbeschlus…  5780 1         Zürich      10     Obfelden  101             
#> # … with 6,509 more rows, and 9 more variables: gebietAusgezaehlt <lgl>,
#> #   jaStimmenInProzent <dbl>, jaStimmenAbsolut <int>, neinStimmenAbsolut <int>,
#> #   stimmbeteiligungInProzent <dbl>, eingelegteStimmzettel <int>,
#> #   anzahlStimmberechtigte <int>, gueltigeStimmen <int>, votedate <date>