Obtain similarities a vote result shares with other votes

similar_votes allows to obtain correlations of specified vote with other votes.

similar_votes(federalvotes = NULL, id = NULL, corr = T,
  from = NULL, to = NULL)

Arguments

federalvotes
tibble or data.frame that is returned by 'get_swissvotes'.

id
identification number of the vote, needs four digits. Vote 626 (Zersiedelungsinitiative) needs 6260.

corr
set to TRUE by default. If FALSE return the variance-covariance matrix.

from
lower limit of correlations.

to
upper limit of correlations.

Value

a tibble containing the results

Examples

 # \donttest{
 federalvotes <- get_swissvotes(geolevel = "canton",from_date = "2010-03-07",to_date="2019-02-10")

#> Error in get_swissvotes(geolevel = "canton", from_date = "2010-03-07",     to_date = "2019-02-10"): unused arguments (geolevel = "canton", from_date = "2010-03-07", to_date = "2019-02-10")

 #Find correlating votes for the 'Zersiedelungsinitiative', 2019-02-10
 results <- similar_votes(federalvotes, id=6260)

#> Error in similar_votes(federalvotes, id = 6260): object 'federalvotes' not found

#Zersiedelungsinitiative, 2019-02-10
 results <- similar_votes(federalvotes, id=6260, from = 0.5)

#> Error in similar_votes(federalvotes, id = 6260, from = 0.5): object 'federalvotes' not found

 OR

#> Error in eval(expr, envir, enclos): object 'OR' not found

#Zersiedelungsinitiative, 2019-02-10
 results <- similar_votes(federalvotes, id=6260, from = 0.1, to = 0.2)

#> Error in similar_votes(federalvotes, id = 6260, from = 0.1, to = 0.2): object 'federalvotes' not found

glimpse(results)

#> Error in glimpse(results): could not find function "glimpse"

# }